+++ packages/core/src/client/types.ts
@@ -8,7 +8,10 @@ import type {
   PublicClient,
   Transport,
 } from "viem";
-import type { PublicRequests } from "viem/dist/types/types/eip1193";
+import type {
+  EIP1193RequestFn,
+  PublicRpcSchema,
+} from "viem/dist/types/types/eip1193";
 import type {
   BigNumberish,
   UserOperationEstimateGasResponse,
@@ -19,32 +22,38 @@ import type {
 
 export type SupportedTransports = Transport | FallbackTransport | HttpTransport;
 
-export type Erc4337Requests = {
-  request(args: {
-    method: "eth_sendUserOperation";
-    params: [UserOperationRequest, Address];
-  }): Promise<Hash>;
-  request(args: {
-    method: "eth_estimateUserOperationGas";
-    params: [UserOperationRequest, Address];
-  }): Promise<UserOperationEstimateGasResponse>;
-  request(args: {
-    method: "eth_getUserOperationReceipt";
-    params: [Hash];
-  }): Promise<UserOperationReceipt>;
-  request(args: {
-    method: "eth_getUserOperationByHash";
-    params: [Hash];
-  }): Promise<UserOperationResponse>;
-  request(args: {
-    method: "eth_supportedEntryPoints";
-    params: [];
-  }): Promise<Address[]>;
-  request(args: {
-    method: "eth_maxPriorityFeePerGas";
-    params: [];
-  }): Promise<BigNumberish>;
-};
+export type Erc337RpcSchema = [
+  {
+    Method: "eth_sendUserOperation";
+    Parameters: [UserOperationRequest, Address];
+    ReturnType: Hash;
+  },
+  {
+    Method: "eth_estimateUserOperationGas";
+    Parameters: [UserOperationRequest, Address];
+    ReturnType: UserOperationEstimateGasResponse;
+  },
+  {
+    Method: "eth_getUserOperationReceipt";
+    Parameters: [Hash];
+    ReturnType: UserOperationReceipt;
+  },
+  {
+    Method: "eth_getUserOperationByHash";
+    Parameters: [Hash];
+    ReturnType: UserOperationResponse;
+  },
+  {
+    Method: "eth_supportedEntryPoints";
+    Parameters: [];
+    ReturnType: Address[];
+  },
+  {
+    Method: "eth_maxPriorityFeePerGas";
+    Parameters: [];
+    ReturnType: BigNumberish;
+  }
+];
 
 export interface Erc4337Actions {
   /**
@@ -98,7 +107,7 @@ export interface Erc4337Actions {
 export interface PublicErc4337Client<T extends SupportedTransports = Transport>
   extends PublicClient<T, Chain>,
     Erc4337Actions {
-  request: (PublicRequests & Erc4337Requests)["request"];
+  request: EIP1193RequestFn<[PublicRpcSchema[number], Erc337RpcSchema[number]]>;
 
   // below methods are not all erc4337 methods, but are the methods we need in the SmartContractAccountProvideer
   getMaxPriorityFeePerGas(): Promise<BigNumberish>;