+++ packages/core/src/signer/private-key.ts
@@ -1,35 +1,35 @@
-import type {Address} from "abitype";
-import type {Hex, PrivateKeyAccount} from "viem";
-import {privateKeyToAccount} from "viem/accounts";
-import type {SmartAccountSigner} from "./types";
+import type { Address } from "abitype";
+import type { Hex, PrivateKeyAccount } from "viem";
+import { privateKeyToAccount } from "viem/accounts";
+import type { SmartAccountSigner } from "./types";
 
 export class PrivateKeySigner implements SmartAccountSigner {
-    owner: PrivateKeyAccount
+  owner: PrivateKeyAccount;
 
-    constructor(owner: PrivateKeyAccount) {
-        this.owner = owner;
-    }
+  constructor(owner: PrivateKeyAccount) {
+    this.owner = owner;
+  }
 
-    signMessage(msg: Uint8Array | Hex | string): Promise<Hex> {
-        if (typeof msg === "string") {
-            return this.owner.signMessage({
-                message: msg
-            })
-        } else {
-            return this.owner.signMessage({
-                message: {
-                    raw: msg
-                }
-            })
-        }
+  signMessage(msg: Uint8Array | Hex | string): Promise<Hex> {
+    if (typeof msg === "string") {
+      return this.owner.signMessage({
+        message: msg,
+      });
+    } else {
+      return this.owner.signMessage({
+        message: {
+          raw: msg,
+        },
+      });
     }
+  }
 
-    getAddress(): Promise<Address> {
-        return Promise.resolve(this.owner.address);
-    }
+  getAddress(): Promise<Address> {
+    return Promise.resolve(this.owner.address);
+  }
 
-    static privateKeyToAccountSigner(key: Hex): PrivateKeySigner {
-        const owner = privateKeyToAccount(key)
-        return new PrivateKeySigner(owner)
-    }
+  static privateKeyToAccountSigner(key: Hex): PrivateKeySigner {
+    const owner = privateKeyToAccount(key);
+    return new PrivateKeySigner(owner);
+  }
 }